Fixing the bug int the sixth task

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 
-income = float(input('Enter your monthly income: '))
-expenses = float(input('Enter your total monthly expenses: '))
+income = int(input('Enter your monthly income: '))
+expenses = int(input('Enter your total monthly expenses: '))
 monthlySavings = income - expenses
 print(f'Your monthly savings are ${monthlySavings}')
 projectedSavings = (monthlySavings * 12) + (monthlySavings * 12 * 0.05)
